+++ app/Modules/Agencies/Exceptions/InvalidAgencySyncCursorException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace App\Modules\Agencies\Exceptions;
+
+use RuntimeException;
+
+class InvalidAgencySyncCursorException extends RuntimeException {}